Implement transaction API
The client can create, read, update and delete transactions. The client
can only create transactions if they are authenticated. Reading,
updating and deleting transactions are exclusive to their respective
owners.

The API endpoints are:

/api/transactions/
  For creating individual transactions, or receiving all of the user's
  transactions in a list.

/api/transactions/<id>/
  For reading, updating, deleting individual transactions.

Delta is used, instead of separate inflow and outflow columns, to
simplify things on the back end.

Furthermore, the API is linked. The client can follow a url from the
list to a specific transaction.

Closes #1