Implement django-allauth
Django-allauth is added to this project. Currently, only local accounts
are enabled, but this might be subject to change later. Only the email
address is requested during registration, no username is required any
more. Furthermore, `/app/transactions` is only accessible now for logged
in users.

Logging in and logging out already had REST API endpoints via
django-rest-framework, but registering can only be done via the web
interface. API endpoints need to be added for this. See #19.

Closes #13